misc/HtmlTestSuite/results/test-0038.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>IndexedDB: Verify [SameObject] behavior of the global scope's indexedDB attribute</title>
<meta href="https://w3c.github.io/IndexedDB/#dom-windoworworkerglobalscope-indexeddb" name="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script>

test(t =&gt; {
  assert_equals(self.indexedDB, self.indexedDB,
                'Attribute should yield the same object each time');

}, 'indexedDB is [SameObject]');

</script>
</head><body/></html>
